# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "初始化路由",
        "title": "API",
        "contact": {
            "name": "wty92911",
            "url": "https://github.com/wty92911",
            "email": "2066946886@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/categories": {
            "get": {
                "description": "获取所有分类的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "获得所有分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "家庭ID",
                        "name": "family_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Category"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/category": {
            "post": {
                "description": "创建分类,包括分类的三级名称、图片，返回创建好的分类+图片链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "创建分类",
                "parameters": [
                    {
                        "description": "家庭ID",
                        "name": "family_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "顶级分类名称",
                        "name": "top_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "中间分类名称",
                        "name": "mid_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "分类名称",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "图片链接",
                        "name": "image_url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Category"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/category/{id}": {
            "delete": {
                "description": "根据分类ID删除分类",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "删除分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gin.H"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "github_com_wty92911_GoPigKit_internal_model.Category": {
            "type": "object",
            "required": [
                "family_id",
                "image_url",
                "mid_name",
                "name",
                "top_name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "family_id": {
                    "type": "integer"
                },
                "foods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.Food"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "mid_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "top_name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wty92911_GoPigKit_internal_model.Food": {
            "type": "object",
            "required": [
                "category_id",
                "created_by",
                "image_urls",
                "price",
                "title"
            ],
            "properties": {
                "category_id": {
                    "description": "食品分类ID，外键",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdUser": {
                    "$ref": "#/definitions/github_com_wty92911_GoPigKit_internal_model.User"
                },
                "created_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "desc": {
                    "description": "食品描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_urls": {
                    "description": "食品图片",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "description": "食品价格",
                    "type": "integer"
                },
                "title": {
                    "description": "食品名称",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wty92911_GoPigKit_internal_model.User": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "family_id": {
                    "description": "user创建时可以不指定family_id，等后面再加入一个家庭",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "open_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}